chore: change errors treatments

+++ apps/backend/src/modules/chat-bot/chat-bot.controller.ts
@@ -69,8 +69,12 @@ export class ChatBotController {
   findManyCompanyChatBots(
     @Query() filterCompanyChatbotsDto: FilterCompanyChatBotsDto,
     @Query() paginationDto: PaginationDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to list the chatbots.'
+      );
     return this.chatBotService.findManyCompanyChatBots(
       filterCompanyChatbotsDto,
       paginationDto,
@@ -92,8 +96,12 @@ export class ChatBotController {
   async getTrainingDataContent(
     @Res() res: Response,
     @Param() { id }: MongoIdDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to get training data.'
+      );
     const fileStream = await this.chatBotService.getTrainingDataContent(
       id,
       userId
@@ -225,7 +233,14 @@ export class ChatBotController {
     status: HttpStatus.NO_CONTENT,
     description: 'Chatbot successfully deleted.',
   })
-  deleteChatBot(@Param() { id }: MongoIdDto, @DBUser() { id: userId }: User) {
+  deleteChatBot(
+    @Param() { id }: MongoIdDto,
+    @DBUser() { id: userId, accountType }: User
+  ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to delete a chatbot.'
+      );
     return this.chatBotService.delete(id, userId);
   }
 
@@ -243,8 +258,12 @@ export class ChatBotController {
   })
   findManyTrainingData(
     @Query() paginationDto: PaginationDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to access the training data.'
+      );
     return this.chatBotService.findManyTrainingData(paginationDto, userId);
   }
 
@@ -262,8 +281,12 @@ export class ChatBotController {
   })
   initChat(
     @Param() { id: chatBotId }: MongoIdDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'PHYSICAL_PERSON')
+      throw new UnauthorizedException(
+        'Only physical person accounts are allowed to initiate a chat.'
+      );
     return this.chatBotService.initChatBot(userId, chatBotId);
   }
 
@@ -281,8 +304,12 @@ export class ChatBotController {
   })
   inviteUser(
     @Body() inviteToChatBotDto: InviteToChatBotDto,
-    @DBUser() { id: companyId }: User
+    @DBUser() { id: companyId, accountType }: User
   ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to send invites.'
+      );
     return this.chatBotService.inviteUser(inviteToChatBotDto, companyId);
   }
 
@@ -299,8 +326,12 @@ export class ChatBotController {
   })
   deleteTrainingData(
     @Param() { id }: MongoIdDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to remove the training data.'
+      );
     return this.chatBotService.deleteTrainingData(id, userId);
   }
 
@@ -317,8 +348,12 @@ export class ChatBotController {
   })
   removeInvitation(
     @Param() { id: invitationId }: MongoIdDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'COMPANY')
+      throw new UnauthorizedException(
+        'Only company accounts are allowed to remove the invitation.'
+      );
     return this.chatBotService.removeInvitation(invitationId, userId);
   }
 
@@ -336,8 +371,12 @@ export class ChatBotController {
   })
   createChatCompletion(
     @Body() createChatCompletionDto: CreateChatBotCompletionDto,
-    @DBUser() { id: userId }: User
+    @DBUser() { id: userId, accountType }: User
   ) {
+    if (accountType !== 'PHYSICAL_PERSON')
+      throw new UnauthorizedException(
+        'Only physical person accounts are allowed to use the chatbot.'
+      );
     return this.chatBotService.createChatCompletion(
       createChatCompletionDto,
       userId
@@ -356,7 +395,14 @@ export class ChatBotController {
     description: 'Chatbot history successfully retrieved.',
     type: GetChatBotHistoryResponseEntity,
   })
-  getChatHistory(@Param() { id }: MongoIdDto, @DBUser() { id: userId }: User) {
+  getChatHistory(
+    @Param() { id }: MongoIdDto,
+    @DBUser() { id: userId, accountType }: User
+  ) {
+    if (accountType !== 'PHYSICAL_PERSON')
+      throw new UnauthorizedException(
+        'Only physical person accounts are allowed to access the chatbot history.'
+      );
     return this.chatBotService.getChatHistory(id, userId);
   }
 }
